pyproject.toml
[project]
name = "data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.1",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.9",
    "openai>=1.0.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.3.3",
    "bs4>=0.0.2",
    "serpapi>=0.1.5",
    "google-search-results>=2.4.2",
    "firecrawl-py>=1.5.0",
    "requests>=2.31.0",
]
